Split media-playback types into one-file-per-type (issue 5409)
Honor the repo's one-major-type-per-file rule for new app-target files: split
BrowserMediaPlayback.swift into BrowserMediaPlaybackReport.swift (struct),
BrowserMediaPlaybackMessageHandler.swift (handler), and the
BrowserPanel+MediaPlayback.swift extension. No behavior change.
