+++ Sources/Panels/BrowserMediaPlaybackReport.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+/// A per-frame media-playback report from the injected media-playback hook.
+struct BrowserMediaPlaybackReport {
+    /// Stable id for the reporting frame's document, so the native side can
+    /// aggregate playback across the main frame and any (cross-origin) iframes.
+    let frameID: String
+    /// Whether that frame currently has any actively-playing media.
+    let isPlaying: Bool
+}
