+++ Sources/Panels/BrowserPanel+MediaPlayback.swift
@@ -1,47 +1,8 @@
 import Foundation
 import WebKit
 
-// MARK: - Bridge contract
-
 /// Name of the `WKScriptMessageHandler` the injected media-playback hook posts to.
 private let mediaPlaybackMessageHandlerName = "cmuxMediaPlayback"
-
-/// A per-frame media-playback report from the injected hook.
-struct BrowserMediaPlaybackReport {
-    /// Stable id for the reporting frame's document, so the native side can
-    /// aggregate playback across the main frame and any (cross-origin) iframes.
-    let frameID: String
-    /// Whether that frame currently has any actively-playing media.
-    let isPlaying: Bool
-}
-
-/// Receives `{ frameID, playing }` from the injected media-playback hook and
-/// forwards it to the owning ``BrowserPanel`` on the main actor.
-///
-/// Mirrors ``ReactGrabMessageHandler``: a thin `NSObject` adapter so the panel
-/// itself never has to conform to `WKScriptMessageHandler`.
-final class BrowserMediaPlaybackMessageHandler: NSObject, WKScriptMessageHandler {
-    private let onReport: @MainActor (BrowserMediaPlaybackReport) -> Void
-
-    init(onReport: @escaping @MainActor (BrowserMediaPlaybackReport) -> Void) {
-        self.onReport = onReport
-    }
-
-    func userContentController(
-        _ userContentController: WKUserContentController,
-        didReceive message: WKScriptMessage
-    ) {
-        guard let body = message.body as? [String: Any],
-              let frameID = body["frameID"] as? String,
-              let playing = body["playing"] as? Bool else { return }
-        let report = BrowserMediaPlaybackReport(frameID: frameID, isPlaying: playing)
-        Task { @MainActor in
-            onReport(report)
-        }
-    }
-}
-
-// MARK: - BrowserPanel extension
 
 extension BrowserPanel {
     /// Injected document-start hook that reports whether the current frame has
